GODRIVER-1530 Fix the enterprise auth test runner.
